refactor: split src/handlers/bash.ts into focused modules (#177)
bash.ts was 699 lines with four functions over 50 lines. Extracted into:
- bash-shared.ts — extractStdout/Stderr/Command helpers + four constants
- bash-git.ts — gitDiffHandler, gitLogHandler, gitStatusHandler
- bash-test.ts — testRunnerHandler
- bash-docker.ts — dockerPsHandler

bash.ts now holds terraformPlanHandler, packageInstallHandler,
buildToolHandler, ghHandler, and the getBashHandler dispatcher (~300 lines).
All exports remain on bash.ts via re-exports so no import sites changed.
